% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_estimates.R
\name{combine_estimates_stan}
\alias{combine_estimates_stan}
\title{Bayesian combination of several, possibly contradictory estimates}
\usage{
combine_estimates_stan(
  yhat,
  sigma,
  conf,
  prior_mu,
  prior_tau,
  low = -Inf,
  up = Inf,
  multi = 0.01,
  flat_prior = FALSE,
  ...
)
}
\arguments{
\item{yhat}{A vector of estimates}

\item{sigma}{A vector of standard errors for each estimates}

\item{conf}{A vector of study confidences}

\item{prior_mu}{Prior mean}

\item{prior_tau}{Prior standard deviation}

\item{low}{Lower bound}

\item{up}{upper bound}

\item{multi}{tau multiplier}

\item{flat_prior}{should a flat prior be used}

\item{...}{Additional parameters passed to rstan::sampling}
}
\description{
Bayesian combination of several, possibly contradictory estimates
}
\examples{
y <- c(1,2,3)
se <- c(2,2,2)
conf <- c(1,1,.5)
combine_estimates_stan(y, se, conf, 1, 5)
}
