% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_estimates.R
\name{rsq_lambda}
\alias{rsq_lambda}
\title{Calculate r-squared and pooling factor for the model}
\usage{
rsq_lambda(y, sigma, theta, tau)
}
\arguments{
\item{y}{Estimates}

\item{sigma}{Standard errors of estimates}

\item{theta}{Posterior samples of theta}

\item{tau}{Posterior samples of tau}
}
\description{
Calculate r-squared and pooling factor for the model
}
