% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangulator-package.R
\docType{package}
\name{triangulator-package}
\alias{triangulator-package}
\alias{triangulator}
\title{Derive a consensus estimate from multiple, possibly contradictory, empirical estimates.}
\description{
The Triangulator is a Shiny user interface designed to help derive consensus estimates of a population quantity (e.g. a population size, a proportion, a mean, etc.) from multiple empirical estimates.
}
